fix: moved some more scripts
