+++ check_archived_drafts.py
@@ -33,7 +33,7 @@ from create_config import create_config
 from extractors.draft_extractor import DraftExtractor
 from job_log import job_log
 from message_factory.message_factory import MessageFactory
-from remove_directory_content import remove_directory_content
+from utility.utility import remove_directory_content
 
 file_basename = os.path.basename(__file__)
