+++ extractors/rfc_extractor.py
@@ -23,7 +23,7 @@ import shutil
 
 from xym import xym
 
-from extract_elem import extract_elem
+from extractors.extract_elem import extract_elem
 from extractors.helper import check_after_xym_extraction, invert_yang_modules_dict, remove_invalid_files
