+++ modules_compilation/file_hasher.py
